Inicializo las celdas en el tablero y los organizo en filas y columnas

+++ cell.h
@@ -1,5 +1,5 @@
 
 typedef struct {
-    short int original_number;
-    short int number;
+    int original_number;
+    int number;
 } cell_t;
